Modify controller to check for deprecation status of runtime.
Adds bit to Exec to indicate action my not longer be activated or updated.
Adds checks for unsupported runtime in Actions API and report proper error.

Drop Swift 2 support
Adds tests.
